requirements.txt
altair==5.5.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
dotenv==0.9.9
executing==2.2.0
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
grad-cam==1.5.5
h5py==3.14.0
hdbscan==0.8.40
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
lightning-utilities==0.14.0
llvmlite==0.44.0
lmdb==1.7.3
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==2.1.2
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
opencv-python==4.11.0.86
openslide-bin==4.0.0.6
openslide-python==1.4.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
patchify==0.2.3
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyWavelets==1.8.0
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shapely==2.0.7
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.48.1
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.3.13
toml==0.10.2
torch==2.6.0
torchmetrics==1.6.3
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
ttach==0.0.3
typing_extensions==4.12.2
tzdata==2025.1
umap==0.1.1
umap-learn==0.5.9.post2
urllib3==2.5.0
wcwidth==0.2.13
